requirements.txt
aiohttp==3.6.2
amqp==5.0.5
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.3.1
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
billiard==3.6.3.0
cached-property==1.5.2
celery==5.0.5
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3
defusedxml==0.6.0
Django==3.0.8
django-appconf==1.0.4
django-cleanup==5.1.0
django-cors-headers==3.5.0
django-dbbackup==3.3.0
django-environ==0.4.5
django-filter==2.3.0
django-imagekit==4.0.2
django-jalali-date==0.3.1
django-mysql==3.8.1
django-redis==4.12.1
django-tinymce==3.0.2
djangorestframework==3.11.1
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.6.0
drf-nested-routers==0.92.5
drf-yasg==1.17.1
freezegun==1.0.0
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
isodate==0.6.0
itypes==1.2.0
jdatetime==3.6.2
Jinja2==2.11.2
kombu==5.0.2
lxml==4.6.2
MarkupSafe==1.1.1
model-bakery==1.2.1
multidict==4.7.6
mysql==0.0.2
mysqlclient==2.0.3
packaging==20.4
pilkit==2.0
Pillow==8.0.0
prompt-toolkit==3.0.16
PyJWT==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-gettext==4.0
pytz==2021.1
redis==3.5.3
requests==2.24.0
requests-toolbelt==0.9.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
soap==0.0.1rc0
sqlparse==0.4.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
yarl==1.6.3
zeep==3.4.0
gunicorn==20.0.4
python-gettext==4.0
